requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
tokenizers>=0.15.0
gradio>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
accelerate>=0.24.0
safetensors>=0.4.0 